Save/load program state + various tweaks
- improved FileOperations.deleteFile()
- specified imports for test classes
- added class doc for InstantJob and RecurringJob
- added XMLElement tags for file export
- TODO: maybe move XML methods to their own class???

+++ src/main/java/FileOperations.java
@@ -3,8 +3,6 @@ package main.java;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.nio.file.StandardCopyOption;
 
 /**
@@ -26,9 +24,12 @@ public class FileOperations {
 		return success;
 	}
 	
-	// Do we even need this?
 	public boolean deleteFile(String path){
-		boolean success = (new File(path)).delete();
+		File temp = new File(path);
+		boolean success = false;
+		if (temp.exists()){
+			success = (new File(path)).delete();
+		}
 		return success;
 	}
